% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{api_camara}
\alias{api_camara}
\title{Wraps an access to the camara API given a relative path and query arguments.}
\usage{
api_camara(path = NULL, query = NULL, asList = FALSE)
}
\arguments{
\item{path}{URL relative to the API base URL}

\item{query}{Query parameters}
}
\description{
Wraps an access to the camara API given a relative path and query arguments.
}
