% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{assertTableIsValid}
\alias{assertTableIsValid}
\title{Garantees that the dataframe x has all the columns passed by y.}
\usage{
assertTableIsValid(x, y, warning = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{y}{vector of characters containing the names of columns.}

\item{warning}{boolean to show the prints}
}
\description{
Garantees that the dataframe x has all the columns passed by y.
}
