% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{coerceTableParams}
\alias{coerceTableParams}
\title{Garantees that the dataframe obj has all the correct types passed by types.}
\usage{
coerceTableParams(obj, types, order_cols = TRUE)
}
\arguments{
\item{obj}{dataframe}

\item{types}{named vector of the columns names and types}
}
\description{
Garantees that the dataframe obj has all the correct types passed by types.
}
