% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadMapasDeVotacao.R
\name{delete_file}
\alias{delete_file}
\title{Deleta um arquivo}
\usage{
delete_file(filepath)
}
\arguments{
\item{filepath}{Caminho do arquivo a ser removido.}
}
\description{
A partir do caminho de um arquivo, deleta-o do computador.
}
