% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{.auxiliary_fetch_autor_camara}
\alias{.auxiliary_fetch_autor_camara}
\title{Retrieves details about an author of a proposition}
\usage{
.auxiliary_fetch_autor_camara(uri)
}
\arguments{
\item{uri}{URL relative to the Deputy url}
}
\value{
A dataframe containing details about the author of a proposition
}
\description{
Fetches a dataframe containing detailed information about the author of the proposition
}
\examples{
\dontrun{
.auxiliary_fetch_autor_camara('https://dadosabertos.camara.leg.br/api/v2/deputados/178854')
}
}
