% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadMapasDeVotacao.R
\name{download_pdf}
\alias{download_pdf}
\title{Baixa um PDF a partir de uma url e um caminho de destino}
\usage{
download_pdf(url, dest_path = "votacao_senado.pdf")
}
\arguments{
\item{url}{URL da requisição}

\item{dest_path}{Caminho + nome do arquivo PDF que será baixado.}
}
\description{
A partir de uma url e de o caminho de destino + nome para o pdf, baixa e salva este arquivo
}
