% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNumberUsingRegex.R
\name{extractDateUsingRegex}
\alias{extractDateUsingRegex}
\title{Extrai data no formato "dd/mm/yyyy" de uma tag dentro de um xml_nodeset html}
\usage{
extractDateUsingRegex(x, tag)
}
\arguments{
\item{x}{xml_nodeset contendo a tag a ser extraída}

\item{tag}{tag html que possui a data}
}
\value{
Data extraída no formato "yyyy-mm-dd"
}
\description{
A partir de um nó xml_nodeset, extrai o texto da tag e recupera uma data no formato "dd/mm/yyyy"
}
\examples{
\dontrun{
extractDateUsingRegex(x, tag = "caption")
}
}
