% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNumberUsingRegex.R
\name{extractNumberUsingRegex}
\alias{extractNumberUsingRegex}
\title{Extrai somente os números de um texto extraído de regex}
\usage{
extractNumberUsingRegex(text, text_regex)
}
\arguments{
\item{text}{Texto onde o regex será aplicado}

\item{text_regex}{Expressão regular onde o texto será extraído para depois serem retornados apenas os números}
}
\value{
Números existentes em um texto extraído a partir de uma expressão regular
}
\description{
A partir de uma expressão regular, extrai o texto desse regex e depois retorna apenas os números existentes
}
\examples{
\dontrun{
extractNumberUsingRegex("p_cod_materia_i=[\\\\d]*&")
}
}
