% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRollcallVotes.R
\name{extractRollcallVotes}
\alias{extractRollcallVotes}
\title{Extrai informações de votos dos senadores a partir de um conjutno de votações}
\usage{
extractRollcallVotes(votacoes)
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir de um dataframe de votações, extrai os dados de votos dos senadores.
}
\examples{
\dontrun{
votacoes = fetchRollcallVotesByInterval(initial_date = "01/02/2019")
 
rollcalls = extractRollcallVotes(votacoes)
}
}
