% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadMapasDeVotacao.R
\name{fetch_all_votos_senado}
\alias{fetch_all_votos_senado}
\title{Extrai informações de votos dos senadores a partir de um conjutno de votações}
\usage{
fetch_all_votos_senado(url_proposicoes = NULL)
}
\arguments{
\item{votacoes_senado_filepath}{Caminho para o csv das votações}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir de um dataframe de votações, extrai os dados de votos dos senadores.
}
