% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadMapasDeVotacao.R
\name{fetch_votos_por_link_votacao_senado}
\alias{fetch_votos_por_link_votacao_senado}
\title{Extrai informações de votos dos senadores a partir de uma url}
\usage{
fetch_votos_por_link_votacao_senado(url, keep = TRUE)
}
\arguments{
\item{url}{URL da requisição}
}
\value{
Dataframe com informações de votos dos senadores
\dontrun{
"https://www25.senado.leg.br:443/web/atividade/materias/-/materia/123903/votacoes#votacao_5952"

etch_votos_por_link_votacao_senado("http://rl.senado.gov.br/reports/rwservlet?legis&report=/forms/parlam/vono_r01.RDF&paramform=no&p_cod_materia_i=123903&p_cod_materia_f=123903&p_cod_sessao_votacao_i=5952&p_cod_sessao_votacao_f=5952")

}
}
\description{
A partir de uma url, extrai os dados de votos dos senadores.
}
