% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchCamaraAuthors}
\alias{fetchCamaraAuthors}
\title{Fetches proposition's authors}
\usage{
fetchCamaraAuthors(idProposicao = NULL)
}
\arguments{
\item{idProposicao}{Proposition's ID}
}
\description{
Fetches a dataframe containing basic information about the authors of the proposition
}
