% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchCamaraProposal}
\alias{fetchCamaraProposal}
\title{Baixa dados sobre uma proposição}
\usage{
fetchCamaraProposal(id)
}
\arguments{
\item{id}{Um ou mais IDs de proposições}
}
\value{
Dataframe
}
\description{
Retorna um dataframe contendo dados sobre uma proposição
}
