% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchCanaraProposals}
\alias{fetchCanaraProposals}
\title{Fetch proposals from Camara dos Deputados by year}
\usage{
fetchCanaraProposals(ano)
}
\arguments{
\item{ano}{The year of the legislative session}
}
\description{
The function allows for retrieving proposals from Camara dos Deputados API by year.
}
\examples{
\dontrun{
library(purrr)
library(progress)
library(data.table)

y=2001:2019

barra <- progress_bar$new(total = length(y),

propostas <- map_df(y, ~{
        barra$tick() 
        fetchCanaraProposals(.x)})
}
 
}
