% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchBlockLeaders.R
\name{fetchGovBlocLeader}
\alias{fetchGovBlocLeader}
\title{Fetch the government's lider ID in the Senate}
\usage{
fetchGovBlocLeader()
}
\description{
Fetch the government's lider ID in the Senate
}
\examples{
\dontrun{
lider = fetchGovBlocLeader()
}

}
