% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchRollcallVotes.R
\name{fetchRollcallVotes}
\alias{fetchRollcallVotes}
\title{Recupera informações das votações nominais do plenário de uma proposição}
\usage{
fetchRollcallVotes(bill_id, year = NULL)
}
\arguments{
\item{bill_id}{Id da proposição}

\item{year}{Ano que filtra as proposições. Caso seja null, todas as votações serão retornadas}
}
\value{
Votações de uma proposição em um ano ou todos, caso nenhum ano seja passado como parâmetro
}
\description{
A partir do id de uma proposição, recupera dados de
votações nominais de plenário que aconteceram na Senado Federal
}
\examples{
\dontrun{
fetchRollcallVotes(bill_id = 135251)
}
}
