% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchRollcallVotesByInterval.R
\name{fetchRollcallVotesByInterval}
\alias{fetchRollcallVotesByInterval}
\title{Recupera informações das votações nominais do plenário em um intervalo de tempo}
\usage{
fetchRollcallVotesByInterval(
  initial_date = "01/02/2019",
  end_date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{initial_date}{Data inicial do período de votações. Formato: "dd/mm/yyyy"}

\item{end_date}{Data final do período de votações. Formato: "dd/mm/yyyy"}
}
\value{
Votações da proposição em um intervalo de tempo
}
\description{
A partir de uma data de início e uma de fim, recupera dados de
votações nominais de plenário que aconteceram na Senado Federal
}
\examples{
\dontrun{
votacoes <- fetchRollcallVotesByInterval(initial_date = "01/02/2019")
}
}
