% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchSenadoProposals}
\alias{fetchSenadoProposals}
\title{Fetch proposals by year from the Senate}
\usage{
fetchSenadoProposals(ano)
}
\arguments{
\item{ano}{A 4 digits integer as 2019.}
}
\value{
Dataframe with few information of a given propsition
}
\description{
Returns a dataframe with few information from a proposal, including id, type, year, name and an indicator if the bill has a final decision or not.
}
\examples{
\dontrun{

df = fetchSenadoProposals(2020)
}
}
