% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSenators.R
\name{fetchSenators}
\alias{fetchSenators}
\title{Fetch a list of actual senators}
\usage{
fetchSenators(what = "atual", query = NULL)
}
\arguments{
\item{what}{a string, if "atual" only the active senators are returned.}
}
\value{
a data.frame
}
\description{
Fetch a list of actual senators
}
\examples{
\dontrun{
dat = fetchSenators()
} 


}
