% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapVotesFromPDF.R
\name{fetchVotesFromRollcallUrl}
\alias{fetchVotesFromRollcallUrl}
\title{Extrai informações de votos dos senadores a partir de uma url}
\usage{
fetchVotesFromRollcallUrl(url, keep = FALSE)
}
\arguments{
\item{url}{URL da requisição}

\item{keep}{if FALSE, the pdf file is deleted}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir de uma url, extrai os dados de votos dos senadores.
}
\examples{
\dontrun{
"https://www25.senado.leg.br:443/web/atividade/materias/-/materia/123903/votacoes#votacao_5952"

fetchVotesFromRollcallUrl("https://rl.senado.gov.br/reports/rwservlet?legis&report=/forms/parlam/vono_r01.RDF&paramform=no&p_cod_materia_i=139005&p_cod_materia_f=139005&p_cod_sessao_votacao_i=6042&p_cod_sessao_votacao_f=6042")
}
}
