% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchYearlyProposals}
\alias{fetchYearlyProposals}
\title{Fetch proposals by year from the Senate}
\usage{
fetchYearlyProposals(ano, casa)
}
\arguments{
\item{ano}{A 4 digits integer as 2019.}
}
\value{
Dataframe with some information of a given propsition
}
\description{
Returns a data frame with few information from a proposal, including id, type, year, name and an indicator if the bill has a final decision or not.
}
\examples{
\dontrun{
fetchYearlyProposals(2020, 'SF'); 
fetchYearlyProposals(2020, 'CD')
}
}
