% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSenatorNameToId.R
\name{matchSenatorNameToId}
\alias{matchSenatorNameToId}
\title{Mapeia um nome eleitoral para id correspondente}
\usage{
matchSenatorNameToId(df)
}
\arguments{
\item{df}{Dataframe a receber o id do parlamentar}
}
\value{
Dataframe target_df contendo coluna id
}
\description{
Recebe dois dataframes contendo nome eleitoral e um deles com informação de id
}
