% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterVotacaoPlenario.R
\name{obterVotacaoPlenarioSenado}
\alias{obterVotacaoPlenarioSenado}
\title{Obtem os dados de votações realizada no plenário do Senado}
\usage{
obterVotacaoPlenarioSenado(
  dataInicio = "20190201",
  dataFim = "20191231",
  query = NULL
)
}
\arguments{
\item{dataInicio}{string para data de início}

\item{dataFim}{string para data fim 
 @examples 
 \dontrun{
dat = obterVotacaoPlenarioSenado(dataInicio = '20190201',   dataFim = '20200831')
 }}
}
\description{
Obtem os dados de votações realizada no plenário do Senado
}
