% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{printWarningAndList}
\alias{printWarningAndList}
\title{Prints a warning and a list.}
\usage{
printWarningAndList(msg, l)
}
\arguments{
\item{msg}{warning message}

\item{l}{list}
}
\description{
Prints a warning and a list.
}
