% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadMapasDeVotacao.R
\name{scrap_votos_from_pdf_senado}
\alias{scrap_votos_from_pdf_senado}
\title{Raspa os dados de votos de um pdf}
\usage{
scrap_votos_from_pdf_senado(pdf_filepath)
}
\arguments{
\item{url}{URL da requisição}

\item{dest_path}{Caminho + nome do arquivo PDF que será baixado.}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir do caminho do pdf, raspa as informações referentes aos votos dos senadores
}
