% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapVotesFromPDF.R
\name{scrapVotesFromPDF}
\alias{scrapVotesFromPDF}
\title{Raspa os dados de votos de um pdf}
\usage{
scrapVotesFromPDF(filepath)
}
\arguments{
\item{filepath}{Caminho + nome do arquivo PDF que será baixado.}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir do caminho do pdf, raspa as informações referentes aos votos dos senadores
}
