% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SenadoBR.R
\docType{package}
\encoding{UTF-8}
\name{SenadoBR}
\alias{SenadoBR}
\title{\Sexpr[stage=build,results=rd]{tools:::Rd_package_title("SenadoBR")}}
\description{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_description("SenadoBR")}
}
\references{
\url{https://dadosabertos.senado.leg.br/}
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SenadoBR")}
}
