% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{switchTableParams}
\alias{switchTableParams}
\title{Returns a conversion function given a type name.}
\usage{
switchTableParams(x)
}
\arguments{
\item{x}{type name}
}
\description{
Returns a conversion function given a type name.
}
