% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteToInteger.R
\name{voteIntegerToString}
\alias{voteIntegerToString}
\title{Recupera descrição do voto a partir do código numérico do voto}
\usage{
voteIntegerToString(voto)
}
\arguments{
\item{voto}{Voto para descrição}
}
\value{
Descrição do voto apssado como parâmetro
}
\description{
Recebe um valor numérico que representa o código do voto e retorna a descrição do mesmo
}
\examples{
\dontrun{
voteIntegerToString(2)
}
}
