% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteToInteger.R
\name{voteStringToInteger}
\alias{voteStringToInteger}
\title{Transforma a string de votos para código numérico}
\usage{
voteStringToInteger(df)
}
\arguments{
\item{df}{Dataframe com a coluna voto}
}
\value{
Dataframe com coluna voto enumerada
}
\description{
Recebe um dataframe com coluna voto e converte o valor para um número inteiro
}
\examples{
\dontrun{
voteStringToInteger(df)
}
}
